+++ packages/core/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.8.5 (2026-09-22)
+
+### 🩹 Fixes
+
+- **live-preview:** accept payloads without viewports [SPA-5272] ([6db6fa7](https://github.com/contentful/experiences/commit/6db6fa7))
+
 ## 0.8.4 (2026-09-21)
 
 ### 🚀 Features